intersight/requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.0
decorator==5.1.0
Flask==2.0.2
idna==3.3
ipdb==0.13.9
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
python-dotenv==0.19.2
requests==2.26.0
rich==10.15.2
six==1.16.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
